add notification preferences to gql user type
Normal users are unable to perform queries on the Account graphql type
so we are adding it to the User type instead. We are also removing the
notification preferences field from courses so we only have to maintain
this field in one place on the User type.

fixes VICE-601
flag=none

/ ---- ---- \
| Test Plan |
\ ---- ---- /

- Navigate to a course notification preferences page
- Everything should still work

Change-Id: Idedd113b8c58be764767df8dda8c59be3ceb1a16
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242009
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>