Support role attribute on any element in the RCE
"role" was not white-listed as a valid attribute. Though it has been
proposed as a global attribute, I don't think it was ever officially
adopted that way in a standard. However it seems to have been implemented
as a global attribute by browsers.  This change adds it as a white-listed
attribute to any element.
This change will address the issue in the old and new RCEs

closes LS-1164
flag=none

test plan:
  - in the RCE create a table and a link and whatever else you want
  - flip to HTML edit view and add a role. (e.g. role="presentation"
    on the table, role="button" on the anchor)
  - flip back to wysiwyg editor mode
  - flip back to html edit mode
  > expect the role attributes to still be in the html
  - save whatever it is you're editing
  > expect the role attributes to still be in the html on the saved page

Change-Id: Idc99c3792184a72211c655ab45707527365b97aa
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242176
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>