Backfill root_account_id for AttachmentAssociation
Note: attachments is a dependency, but there are attachment records
that can't be filled so will be null until we fill them with a dummy
(0) root account id. This backfill won't be unlocked until then.

Follows the strategy we discussed of using attachment.root_account_id
if context is ConversationMessage (these can have multiple root account
IDs so it's messy)

This should be merged after the work on Favorites to allow cross-shard
associations (an AttachmentAssociation's attachment can be on a
different shard). This is important as that code also enforces the
possible dependency of attachment_associations on cross-shard
attachments tables being backfilled. (Also, by then we will have
hopefully resolved the discrepancy between Attachment#root_account_id
and the attachment's root_account_id in the database)

Also, I'm not sure if context can ever be nil but the spec was giving it
a nil context (a bug in the spec but whatever, checking for a nil
context doesn't hurt)

closes INTEROP-5842
flag=None

Test plan:
* make an attachment for one root account
* make a course, group, submission, conversation message for another root
  account(s)
* create an AttachmentAssociation for the attachment and each of the above
  four contexts
* make sure the course, group, and submission AttachmentAssociations you
  created got their root_account_id from the context, and the
  conversation message AttachmentAssociation got the root_account_id
  from the attachment
* make sure all attachments have a root_account_id (needed for testing
  here until we have the attachments backfill)
* null out root_account_id on all the AttachmentAssociations and run
  DataFixup::PopulateRootAccountIdOnModels.run
* again make sure the course, group, and submission
  AttachmentAssociations got their root_account_id filled in from the
  context, and the conversation message one got its root_account_id from
  the attachment

Change-Id: Ic946d71073a54873bfb493faff2607cd6ce4b108
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243117
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>