Fix media upload being placed in the wrong place
When uploads take a long time to upload, it's possible for
the uploaded media to be inserted in the wrong place if
the cursor changes before the upload is complete. This
patchset stores a bookmark reference. When the upload
is complete and the upload is inserted in the editor, it
jumps back to the saved bookmark, inserts the image
and returns to where the cursor was.

Test plan:
With new RCE enabled
On a page, add several lines of text.
Get an image ready to upload
Open your network dev tools panel and
change the network speed from 'Online' to
'Slow 3G' - this makes it easier to change the
cursor before the upload is complete.
Start the upload
Quickly change the cursor to a different location
in your editor
Wait for the upload to complete.
The image should have inserted at the location
where the cursor was before the upload was started

flag=rce_enhancements

Fixes LS-1233

Change-Id: If5d70989586214f22dbee998fdf655c603ef8ee0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243861
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Alex Anderson <raanderson@instructure.com>