Add Usage Rights option to image upload
This only applies to uploading images, not other media
like videos and files. Those should be handled in a
separate patchset.

This also only appears when the "Require Usage Rights"
course setting is enabled.

When "Require Usage Rights" is enabled, the Usage
Rights accordion should be expanded already. When it
is not enabled, the Attributes accordion should be
expanded.

For some reason, the Session properties were never
being fetched, so this adds some code to fetch the
session when the upload modal mounts. The session
includes the boolean for whether usage rights are
 required.

Test Plan

Turn on RCE Enhancements.
For a course, turn on the "Copyright and license
information must be provided for files before they
are published." setting on the front page of the
course settings.
Open up an RCE.
Open the upload modal. The "Usage Rights" box
should appear beneath where the file upload is.
It should be expanded.
Upload a file and adjust the settings in the
usage rights section.
Go to the Files page and find the file that you
just uploaded.
On the right side of the file's row, it should have
your file's usage rights data listed correctly.

Refs LS-324

flag = rce_enhancements

Change-Id: Ie1f9beccc3d28dc9eb37c59c5166cd13e6fe91af
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/244618
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>