report runner improvements
this doesn't really change anything with how report runners are run, but
how they might fail.
we don't want to abort the runner that may have failed, so we can get
data on which report had an error with state of error or running.

Also we don't want to write rows on a failed or aborted report runner
because then we can manually run that runner and any aborted to finish
the report in a console for troubleshooting.

test plan 
 - specs should pass

Change-Id: I5cfc3d975899f85aff64a6e20c7673b1f2cd1b88
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/244625
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>