Dev key modal: fix deceiving validation error
The validation error under the "Redirect URIs" message was shown if there
is _any_ validation error, whether or redirect URIs are present.

Note that there is also already a flashed message if Redirect URIs is
invalid. I considered just getting rid of the validation error under the
field. I can still do that if people think that is better.

flag=none
refs INTEROP-5821

Test plan:
- Try to create an LTI with no redirect uri. Check that the validation
error message is shown. This behavior is unchanged.
- Add a redirect uri but leave out title. Check that the validation
message for title is shown but not the validation message for redirect
URI. This behavior _is_ changed.
- Try to create an API key with enforced scopes but no scopes. Check
that a validation error message is shown but that the message for
required "Redirect URIs" is not shown [behavior unchanged]

Change-Id: Ie5666dacb95f9dfb92c7b40c931938964f1b813c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245875
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Karl Lloyd <karl@instructure.com>
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>