prevent creating logins when restoring old email
communication_channels can create pseudonym on creation if they are from
the default shard. This is only on Account.default because it is the way
for a user to add a login for FFT, but not enabled on other accounts
because schools will manage the authentication either through 3rd party
or within canvas using canvas auth, but still controlled by admins,
where there're no admins for FFT.

the model changes are not needed, but would ensure that it happened on
the accounts shard instead of the users shard. There is no real change
there.

test plan
 - specs should pass
if you really want to test it in addition to specs, it would be like
 - with user on default account, create a communication channel on
   profile page
 - check the box to be able to authenticate with it.
 - delete the communication channel.
 - re-add the communication channel and don't check the box.
 - confirm the communication channel
 - it should not create a second pseudonym for the user.

fixes VICE-768
flag=none

Change-Id: I8a5c9e18311417c97aabe7c1f13e192ba9c69b1c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/246866
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ben Nelson <bnelson@instructure.com>
QA-Review: Ben Nelson <bnelson@instructure.com>
Product-Review: Ben Nelson <bnelson@instructure.com>