Fix up CanvasDateInput and use it in MutationAuditLog
Refs FOO-625
flag=none

The new InstUI DateInput is way less opinionated than the old
one, so it takes a lot more handholding to make it function.
Fortunately someone had written much of a Canvas driver for
that component and it is only going to need a little tweaking.

As a first test of this, I did the MutationAuditLog requester
using this new component.

Test plan:
* Go to Admin Tools -> Logging -> GraphQL Mutation Activity
* That form doesn't look exactly the same, but should have all
  the same functionality as it used to

Change-Id: Ic61e91623562e7990f861ef44019ba92f5e87fd8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/247425
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>