Load mathjax if we find any inline LaTex
If the LaTex the user enters in the equation editor has errors, MathJax
will typeset it, with the bad parts in red. Unfortunately the services
canvas uses to generate the equation image simply fails. The equation
editor now verifies the image can be created and if not will insert the
raw LaTex into the RCE delimited by \(...\). Once the content is saved
MathJax will typeset it (includeing the red, but hey, that's better than
what used to happen)

now also looks for $$ or \( starting delimiters mathjax uses to identify
equations

and looks for math on wiki pages, which was missing before (that is not
behind the flag)

had to skip some screenreader_gradebook specs that started failing with
this change. The gradebook team is as baffled as I am. Will create a ticket
for them to investigate.

if MathJax finds LaTex errors (which you'll see as red in the preview
area), the service that generates the equation image fails. This detects
that and puts the raw LaTex back into the RCE's content. On saving,
MathJax can process it, though it will still show the red. This is better
than losing the equation because the image was missing.

closes LS-1488
flag=inline_math_everywhere

test plan
  - turn on SiteAdmin feature "Updated math equation handling"
    (flag=new_math_equation_handling) and the RootAccount feature
    "Support LaTex math equations almost everywhere"
    (flag=inline_math_everywhere)
  - in the RCE, enter \(your equation here\)
  - enter $$your equation here$$
  - save
  > expect the first equation to be typeset and inline
  > expect the second equation to be typeset and a block
  - the equations can span over multiple lines too
  - edit an eq with the eq editor
  > expect it to be typeset too

  - in the RCE, click on the equation toolbar button
  - in the advanced pane of the equation editor, enter \var = 27
  > expect the \var part to be red in the preview area (because it's undefined)
  - click Insert Equation
  > expect the raw LaTex to be in the RCE, surrounded by \(...\)
  - save
  > expect the equation to get typeset (and \var will still be red)

  - have an assignment with a student submission
  - grade the submission and include an equation in the comments
    for example: \(y = x^2\)
  - view the assignment as the student
  > expect the equation to be renered in the comment
  - put some math in an assignment's title
  > expect to see it in the assignments page

  - create a page with inline math in it
  > expect the math to be typeset

  - create a discussion with math
  - reply to the discussion with inline math in it
  - edit the inline math in the reply
  > expect the math to be typeset each time

  - have a student submit an assignemnt
  - as the teacher go to speedgrader
  > expect typeset math everywhere
  - add a comment with math
  > expect it to get typeset

  - goto /assignments/:id/submissions/:user_id
  > expect  comments with math to be typeset
  - add a comment with math
  > epxect it go get typeset

  - old quizzes mostly work. While editing a question, inline
    math is not typeset, but will be when previewed or being
    taken

  - new module items can have math in their title, but if you edit
    the title, it's not typeset (this has baffled me, but it's
    got to be an extremely rare case)

  - load the calendar and flipping between Week, Month, Agenda views
  > expect entries with inline math in their title to be typeset
  - loading the calendar in different initial Views
  > expect events with math to be typeset
  - create a new event with math (e.g. "\(x^2\)"in the title
  > expect it to be typeset when saved
  - open an event in the calendar that has math in its body
  > expect it to be typeset

Change-Id: Id5e1e822fad29a52bf21573e62976a4482afcf43
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/248246
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>