skip css:styleguide when python2 is not available
refs FOO-966

pygments.rb, a dependency of dress_code, requires a python 2 interpreter
and will look for it in PATH (under the name of "python2"). This patch
makes the requirement more graceful in development where we skip the
task when python2 is unavailable, and issue a warning with instructions
on how to make it work.

|
| TEST PLAN
|

I used pyenv[1] to install python3, but you can use virtualenv too or
whatever your system has; all you need is access to both python2 and
python3 interpreters.

- with python2 available in PATH, verify the task works fine:

    bundle exec rake css:styleguide

- with python != 2, verify the task is skipped and a warning is issued
  but the process exits with 0 (success):

    PYGMENTS_RB_PYTHON=python3 \
      bundle exec rake css:styleguide

    echo $? # => 0

[1]: https://github.com/pyenv/pyenv

Change-Id: Icfdcfa6a17a879f05e0528b97e8202c5a2db986d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/248436
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>