fix race condition on report runners
this is very difficult to test, because it's a race condition, but the
test plan would work...

the changes are just plumbing through if the primary db should be used
or a replica, and then methods that have been well tested and used.

for parallel account_reports we write rows to account_report_rows and
then read from account_report_rows to generate the csv file. If this is
done on a replica, it can be lagging and not get all the records.
Typical reports, this would not be a problem because it is old data for
a report...
but when we just wrote the data it may not exist, so use the primary
database.

test plan
 - run a report with a primary database and read_only replica
 - pause replication
 - the report should work

fixes PFS-16619
flag = none

Change-Id: I93140ce8f566a4e0553377388d6296ec36d800cb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/248863
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mark Valentine <mvalentine@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>